requirements.txt

python-telegram-bot==21.5
openai>=1.50.0
Pillow
python-dotenv
